backend/requirements.txt
fastapi
uvicorn
supabase
python-multipart
python-dotenv
ffmpeg-python
boto3